requirements.txt
colorama==0.4.4
fpdf==1.7.2
isbnlib==3.10.8
mysql_connector_repackaged==0.3.1
numpy==1.20.1
openpyxl==3.0.7
pandas==1.2.4
pretty_html_table==0.9.16
prettytable==3.5.0
reportlab==3.6.12
SQLAlchemy==1.4.44
tabulate==0.9.0
termcolor==2.1.0
win32printing==0.1.3
xlsxwriter==3.0.3
